+++ conf/urls.py
@@ -12,6 +12,7 @@ import conf.sitemaps
 import euexit.views
 import invest.views
 import contact.views
+import find_a_supplier.views
 
 
 sitemaps = {
@@ -96,6 +97,16 @@ urlpatterns += [
         r'^international/content/invest/$',
         QuerystringRedirectView.as_view(pattern_name='invest-home'),
         name='content-invest-home-redirect'
+    ),
+    url(
+        r'^international/trade/incoming/$',  # Homepage
+        QuerystringRedirectView.as_view(pattern_name='trade-home'),
+        name='trade-incoming-homepage'
+    ),
+    url(
+        r'^international/trade/incoming/(?P<path>[\w\-/]*)/$',
+        find_a_supplier.views.LegacySupplierURLRedirectView.as_view(),
+        name='trade-incoming'
     ),
     url(
         r'^international/trade/$',