+++ find_a_supplier/views.py
@@ -1,0 +1,7 @@
+from core.views import LegacyRedirectCoreView
+from . import redirects
+
+
+class LegacySupplierURLRedirectView(LegacyRedirectCoreView):
+    redirects_mapping = redirects.REDIRECTS
+    fallback_url = '/international/trade/'